requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cmdstanpy==1.2.4
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2

# Dependency pypi:djangorestframework-simplejwt:5.3.1 is vulnerable
# CVE-2024-22513,  Score: 8.8
# The djangorestframework-simplejwt is vulnerable to information disclosure.
# A user can access web application resources even after their account has been disabled
# due to missing user validation checks via the "for_user" method.
# Read More: https://devhub.checkmarx.com/cve-details/CVE-2024-22513?utm_source=jetbrains&utm_medium=referral
# Results powered by Checkmarx ©
djangorestframework-simplejwt==5.3.1

drf-yasg==1.21.7
fonttools==4.53.1
holidays==0.55
idna==3.7
importlib_resources==6.4.3
inflection==0.5.1
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prophet==1.1.5
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
stanio==0.5.1
tqdm==4.66.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
